+++ src/mapgen/mapgen.h
@@ -38,7 +38,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #define MG_DECORATIONS 0x20
 #define MG_BIOMES      0x40
 
-typedef u8 biome_t;  // copy from mg_biome.h to avoid an unnecessary include
+typedef u16 biome_t;  // copy from mg_biome.h to avoid an unnecessary include
 
 class Settings;
 class MMVManip;
